requirements.txt
# Core Dependencies
MetaTrader5>=5.0.45
numpy>=1.24.0
pandas>=2.0.0
polars>=0.18.0
numba>=0.57.0

# ML/AI
torch>=2.0.0
scikit-learn>=1.3.0
hmmlearn>=0.3.0
TA-Lib>=0.4.26

# Performance
asyncio-mqtt>=0.16.1
aiohttp>=3.8.0
uvloop>=0.17.0; platform_system != "Windows"
redis>=4.5.0

# Dashboard
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0.0
python-multipart>=0.0.6

# Data Processing
scipy>=1.11.0
statsmodels>=0.14.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
structlog>=23.1.0
psutil>=5.9.0
colorama>=0.4.6

# Database
aiosqlite>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code Quality
black>=23.7.0
ruff>=0.0.285
mypy>=1.5.0

# Monitoring (optional)
prometheus-client>=0.17.0
